create button component, apply it across pages
