+++ src/app/admin/categories/_components/CategoryForm.tsx
@@ -1,9 +1,8 @@
 'use client'
 import { FieldErrors, UseFormHandleSubmit, UseFormRegister } from "react-hook-form";
 import { CreateCategoryRequestBody } from "@/app/_types/type";
-import { useParams, useRouter } from "next/navigation";
-import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
-import { mutate } from "swr";
+import ErrorMessage from "@/app/_components/ErrorMessage";
+import Button from "../../_components/Button";
 
 type FormValues = {
   name: string;
@@ -19,40 +18,6 @@ interface Props {
 }
 
 const CategoryForm: React.FC<Props> = ({ register, handleSubmit, errors, isSubmitting, submitFunction, mode }) => {
-  const params = useParams();
-  const id = params.id
-  const router = useRouter();
-  const { token } = useSupabaseSession();
-  
-  // DELETE: カテゴリー削除
-  const handleDelete = () => {
-    if (!token) return
-
-    const fetcher = async () => {
-      try {
-        const res = await fetch(`/api/admin/categories/${id}`, {
-          method: 'DELETE',
-          headers: {
-            'Content-Type': 'application/json',
-            authorization: token,
-          },
-        })
-
-        if (!res.ok) {
-          throw new Error();
-        } else {
-          alert('カテゴリーを削除しました');
-          mutate(`/api/admin/categories/${id}`);
-          router.replace('/admin/categories');
-        }
-      } catch (error) {
-        console.error('エラーが発生しました。', error);
-      }
-    }
-
-    fetcher();
-  }
-  
   return (
     <form className='w-[100%] flex flex-col' id='myForm' onSubmit={handleSubmit(submitFunction)}>
       <label className='text-base text-[#4d4f5b]' htmlFor="name">カテゴリー名</label>
@@ -61,16 +26,8 @@ const CategoryForm: React.FC<Props> = ({ register, handleSubmit, errors, isSubmi
           required: 'カテゴリー名を入力して下さい。',
         })}
       />
-      <div>{errors.name?.message ?? ''}</div>
-      <div>
-          <button 
-          className='w-[80px] mt-[10px] pt-2 px-4 pb-[10px] text-[rgb(255,255,255)] 
-          text-lg bg-[#4085fd] border-0 rounded-lg font-bold cursor-pointer' 
-          form='myForm' type="submit">{mode === 'new' ? '作成' : '更新'}</button>
-          {mode === 'edit' ? <button className='w-[80px] mt-[10px] ml-[10px] pt-2 px-4 pb-[10px] 
-          text-[rgb(255,255,255)] text-lg bg-[#fd4d40] border-0 rounded-lg font-bold 
-          cursor-pointer' type='button' onClick={handleDelete}>削除</button> : ''}
-        </div>
+      <ErrorMessage errors={errors} name='name' />
+      <Button mode={mode} page='categories'/>
     </form>
   )
 }
